feat: create new Dropdowns components, fix: api data.data -> results
